+++ clifs/plugins/rename.py
@@ -1,24 +1,26 @@
-"""Clifs plugin for regex-based file renaming"""
+"""Clifs plugin for regex-based renaming of files and folders"""
 
 import re
 import sys
 from argparse import ArgumentParser, Namespace
-from collections import Counter
+
+# from collections import Counter
 from pathlib import Path
-from typing import List, Set
+from typing import Counter, List, Literal, Set
 
 from clifs import ClifsPlugin
 from clifs.utils_cli import cli_bar, print_line, set_style, user_query
-from clifs.utils_fs import INDENT, FileGetterMixin, get_unique_path
+from clifs.utils_fs import INDENT, PathGetterMixin, get_unique_path
 
 
-class FileRenamer(ClifsPlugin, FileGetterMixin):
+class Renamer(ClifsPlugin, PathGetterMixin):
     """
-    Regex-based file renaming.
+    Regex-based renaming of files and folders.
     """
 
     pattern: str
     replacement: str
+    rename_dirs: bool
     skip_preview: bool
 
     @classmethod
@@ -48,6 +50,13 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
             "Defaults to empty string.",
         )
         parser.add_argument(
+            "-d",
+            "--dirs",
+            dest="rename_dirs",
+            action="store_true",
+            help="Rename directories instead of files.",
+        )
+        parser.add_argument(
             "-sp",
             "--skip_preview",
             action="store_true",
@@ -57,40 +66,62 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
 
     def __init__(self, args: Namespace) -> None:
         super().__init__(args)
-        self.files2process: List[Path] = self.get_files()
-        self.counter = Counter(files2process=len(self.files2process))
+
+        self.counter: Counter[str] = Counter()
+        self.files, self.dirs = self.get_paths()
+        if self.rename_dirs:
+            # process deeper folders first to avoid changing paths on the fly
+            self.dirs = self.sort_paths(self.dirs)
 
     def run(self) -> None:
-        self.exit_if_nothing_to_process(self.files2process)
+        if not self.rename_dirs:
+            if not self.files:
+                self.console.print("No files to process")
+            else:
+                if not self.skip_preview:
+                    self.rename(self.files, path_type="files", preview_mode=True)
+                    if not user_query(
+                        'If you want to apply renaming, give me a "yes" or "y" now!'
+                    ):
+                        self.console.print("Will not rename for now. See you soon.")
+                        sys.exit(0)
+                self.rename(self.files, path_type="files", preview_mode=False)
+        else:
+            if not self.dirs:
+                self.console.print("No dirs to process")
+            else:
+                if not self.skip_preview:
+                    self.rename(self.dirs, path_type="dirs", preview_mode=True)
+                    if not user_query(
+                        'If you want to apply renaming, give me a "yes" or "y" now!'
+                    ):
+                        self.console.print("Will not rename for now. See you soon.")
+                        sys.exit(0)
+                self.rename(self.dirs, path_type="dirs", preview_mode=False)
 
-        if not self.skip_preview:
-            self.rename_files(preview_mode=True)
-            if not user_query(
-                'If you want to apply renaming, give me a "yes" or "y" now!'
-            ):
-                self.console.print("Will not rename for now. See you soon.")
-                sys.exit(0)
-
-        self.rename_files(preview_mode=False)
-
-    def rename_files(
+    def rename(
         self,
+        paths: List[Path],
+        path_type: Literal["files", "dirs"],
         preview_mode: bool = True,
     ) -> None:
-        self.console.print(f"Renaming {self.counter['files2process']} files.")
+        self.counter.clear()
+        self.counter["paths_total"] = len(paths)
+
+        self.console.print(f"Renaming {self.counter['paths_total']} {path_type}.")
         print_line(self.console)
-        files_to_be_added: Set[Path] = set()
-        files_to_be_deleted: Set[Path] = set()
+        paths_to_be_added: Set[Path] = set()
+        paths_to_be_deleted: Set[Path] = set()
         if preview_mode:
             self.console.print("Preview:")
 
-        num_file = 0
-        for num_file, path_file in enumerate(self.files2process, 1):
-            name_old = path_file.name
+        num_path = 0
+        for num_path, path in enumerate(paths, 1):
+            name_old = path.name
             name_new = re.sub(self.pattern, self.replacement, name_old)
             message_rename = f"{name_old:35} -> {name_new:35}"
 
-            # skip files if renaming would result in bad characters
+            # skip items if renaming would result in bad characters
             found_bad_chars = self.find_bad_char(name_new)
             if found_bad_chars:
                 message_rename += set_style(
@@ -101,22 +132,22 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
                 self.counter["bad_results"] += 1
                 self.print_rename_message(
                     message_rename,
-                    num_file,
+                    num_path,
                     preview_mode=preview_mode,
                 )
                 continue
 
             # make sure resulting paths are unique
-            path_file_new = path_file.parent / name_new
-            path_file_unique = get_unique_path(
-                path_file_new,
-                set_taken=files_to_be_added,
-                set_free=files_to_be_deleted | {path_file},
+            path_new = path.parent / name_new
+            path_unique = get_unique_path(
+                path_new,
+                set_taken=paths_to_be_added,
+                set_free=paths_to_be_deleted | {path},
             )
 
-            if path_file_new != path_file_unique:
-                path_file_new = path_file_unique
-                name_new = path_file_unique.name
+            if path_new != path_unique:
+                path_new = path_unique
+                name_new = path_unique.name
                 message_rename = f"{name_old:35} -> {name_new:35}"
                 message_rename += set_style(
                     f"{INDENT}Warning: name already exists. Adding number suffix.",
@@ -124,35 +155,35 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
                 )
                 self.counter["name_conflicts"] += 1
 
-            # skip files that are not renamed
-            if path_file_new == path_file:
+            # skip items that are not renamed
+            if path_new == path:
                 message_rename = set_style(message_rename, "bright_black")
                 self.print_rename_message(
                     message_rename,
-                    num_file,
+                    num_path,
                     preview_mode=preview_mode,
                 )
                 continue
 
             self.print_rename_message(
                 message_rename,
-                num_file,
+                num_path,
                 preview_mode=preview_mode,
             )
             if not preview_mode:
-                path_file.rename(path_file_new)
-                self.counter["files_renamed"] += 1
+                path.rename(path_new)
+                self.counter["paths_renamed"] += 1
             else:
-                files_to_be_added.add(path_file_new)
-                if path_file_new in files_to_be_deleted:
-                    files_to_be_deleted.remove(path_file_new)
-                files_to_be_deleted.add(path_file)
+                paths_to_be_added.add(path_new)
+                if path_new in paths_to_be_deleted:
+                    paths_to_be_deleted.remove(path_new)
+                paths_to_be_deleted.add(path)
 
         if self.counter["bad_results"] > 0:
             self.console.print(
                 set_style(
                     f"Warning: {self.counter['bad_results']} out of "
-                    f"{self.counter['files2process']} files not renamed as it would "
+                    f"{self.counter['paths_total']} files not renamed as it would "
                     "result in bad characters.",
                     "warning",
                 )
@@ -162,7 +193,7 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
             self.console.print(
                 set_style(
                     f"Warning: {self.counter['name_conflicts']} out of "
-                    f"{self.counter['files2process']} renamings would have resulted in "
+                    f"{self.counter['paths_total']} renamings would have resulted in "
                     "name conflicts. Added numbering suffixes to get unique names.",
                     "warning",
                 )
@@ -170,8 +201,8 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
 
         if not preview_mode:
             self.console.print(
-                f"Hurray, {num_file} files have been processed, "
-                f"{self.counter['files_renamed']} have been renamed."
+                f"Hurray, {num_path} {path_type} have been processed, "
+                f"{self.counter['paths_renamed']} have been renamed."
             )
         print_line(self.console)
 
@@ -187,7 +218,7 @@ class FileRenamer(ClifsPlugin, FileGetterMixin):
         else:
             cli_bar(
                 num_file,
-                self.counter["files2process"],
+                self.counter["paths_total"],
                 suffix=space_prefix + message,
                 console=self.console,
             )